a) A full one developer license was bought for JTattoo-1.6.11 to support the dark theme in the application. The Application.java was modified to choose either the JTattoo or if missing the default scheme.
b) The TEMAllied was changed to keep track instances of scales

c) The TemAdapterScaleTreatment was updated to respond also on movement
when clicking inside the scale but not to the side grips.  


2 major points need to be done in the next time:

a) Better documentation of the scale treatment 
b) Handling of several scales
c) Refactoring that simplifies the source code ... I have the feeling, 
that the classes get too big. 

The TEMAllied class should delegate to it's members to simplify the 
structure ...