I am soooo happy -- Multiple Scales added !!! But see below ...
There is still some work to do concerning the scales - e.g.
there is a exception thrown, when in Scale mode someone 
tries to click into vacuum to move a scale. But from the
rest it seems not to bad. Next thing to make better:

Get rid of exceptions - maybe implement a contract concerning
null references given back (exception...it is not really one) 
What is a better alternative. 

Secondly we should somehow get the TemAdapterScaleTreatment directly 
bound to the Scale ... can we generalize this, so that other 
Drawables need not to take care about their clicking, moving 
etc features anymore in that detail.