Decoratator pattern was added for Drawables.
Now it is not necessary anymore to define setters and getters on 
Drawables but directly use a builder for decorations. The obtained
Decorations can be applied to the Drawable primitives to set the color,
the filling type, the stroke type. 

Missing till now is a another Builder to be used for Text items. 

A  bug has been fixed, that the information line top left in the 
TEMView was not updated when the TEM picture was chosen from the 
TemTable directly.