# WARNING: head commit changed in the meantime
Small preparing changes ... next thing ... generalize an "Clickable"
interface and implement an abstract class with component Drawable ...
that allows to add these properties to any Drawable. This makes it later
easier to handle in the temView. 

Java 8 is necessary for a correct functioning of the 3D cursor. 
Java 7 seems to be broken in this regard.